requirements.txt
pandas
fastapi
uvicorn
scikit-learn
xgboost
python-multipart
joblib
google-generativeai
python-dotenv
